mycode/Day9/currency_converter/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
pydantic>=1.8.2
python-decouple>=3.4
python-dotenv